refactor(protocol): rename sig flow test file and update test
